pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ephraim"
version = "0.3.0"
description = "Senior-Engineer Terminal Coding Agent"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Ephraim Project"}
]
keywords = ["ai", "coding", "agent", "terminal", "git"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development",
]

dependencies = [
    "ollama>=0.1.0",
    "rich>=13.0.0",
    "requests>=2.28.0",
    "beautifulsoup4>=4.11.0",
    "duckduckgo-search>=4.0.0",
    "prompt-toolkit>=3.0.0",
    "lxml>=4.9.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
multimodal = [
    "pdf2image>=1.16.0",
    "PyMuPDF>=1.23.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]
all = [
    "pdf2image>=1.16.0",
    "PyMuPDF>=1.23.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
Ephraim = "ephraim.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["ephraim*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
